Add entities and repos

+++ src/main/resources/db/migration/V1.0.0__create_tables.sql
@@ -1,17 +1,17 @@
 create table if not exists product (
-    id serial primary key,
+    id bigserial primary key,
     name varchar(64) not null,
     price real not null
 );
 
 create table if not exists coupon (
-    id serial primary key,
+    id bigserial primary key,
     code varchar(8) not null unique,
-    percent smallint not null check ( percent > 0 and percent <= 100)
+    percent integer not null check ( percent > 0 and percent <= 100)
 );
 
 create table if not exists tax (
-    id serial not null,
+    id bigserial not null,
     code_pattern varchar(64) not null unique,
-    percent smallint not null check ( percent > 0 and percent <= 1000)
+    percent integer not null check ( percent > 0 and percent <= 1000)
 );
